requirements.txt
numpy
streamlit
pandas
matplotlib.pyplot
seaborn
